Iron out releaser.yml (#278)
