+++ cmd/arrai/sync_darwin.go
@@ -1,0 +1,13 @@
+// +build darwin
+
+package main
+
+// TODO: Allow watching a single file to represent a non-tuple database.
+
+import (
+	"github.com/urfave/cli/v2"
+)
+
+func sync(c *cli.Context) error {
+	panic("not implemented")
+}
