+++ rel/value_set_native_func.go
@@ -3,6 +3,7 @@ package rel
 import (
 	"fmt"
 	"reflect"
+	"unsafe"
 
 	"github.com/arr-ai/hash"
 	"github.com/arr-ai/wbnf/parser"
@@ -41,7 +42,7 @@ func (f *NativeFunction) Fn() func(Value) Value {
 
 // Hash computes a hash for a NativeFunction.
 func (f *NativeFunction) Hash(seed uintptr) uintptr {
-	return hash.String(f.String(), hash.Uintptr(9714745597188477233, seed))
+	return hash.String(f.String(), hash.Uintptr(9714745597188477233>>(64-8*unsafe.Sizeof(uintptr(0))), seed))
 }
 
 // Equal tests two Values for equality. Any other type returns false.
